+++ tests/integ/modin/frame/test_where.py
@@ -576,7 +576,8 @@ def test_dataframe_where_with_np_array_cond():
     )
 
 
-@sql_count_checker(query_count=1, join_count=2)
+# one extra query to convert index to native pandas when creating the snowpark pandas dataframe
+@sql_count_checker(query_count=2, join_count=2)
 def test_dataframe_where_with_np_array_cond_mismatched_labels():
     data = [1, 2, 3]
     cond = np.array([[False, True, False]]).T
@@ -589,7 +590,7 @@ def test_dataframe_where_with_np_array_cond_mismatched_labels():
 
     native_df = native_pd.DataFrame(data, columns=["A"])
     native_other_df = native_pd.DataFrame(
-        other, columns=["B"], index=pd.Index([1, 2, 3], name="A")
+        other, columns=["B"], index=native_pd.Index([1, 2, 3], name="A")
     )
 
     eval_snowpark_pandas_result(
@@ -601,7 +602,7 @@ def test_dataframe_where_with_np_array_cond_mismatched_labels():
     )
 
 
-@sql_count_checker(query_count=1, join_count=2)
+@sql_count_checker(query_count=2, join_count=2)
 def test_dataframe_where_with_dataframe_cond_single_index_different_names():
     data = [1, 2, 3]
     cond = [False, True, False]
@@ -616,7 +617,7 @@ def test_dataframe_where_with_dataframe_cond_single_index_different_names():
     native_df = native_pd.DataFrame(data, columns=["A"])
     native_cond_df = native_pd.DataFrame(cond, columns=["A"])
     native_other_df = native_pd.DataFrame(
-        other, columns=["B"], index=pd.Index([1, 2, 3], name="A")
+        other, columns=["B"], index=native_pd.Index([1, 2, 3], name="A")
     )
 
     eval_snowpark_pandas_result(
@@ -628,7 +629,8 @@ def test_dataframe_where_with_dataframe_cond_single_index_different_names():
     )
 
 
-@sql_count_checker(query_count=1, join_count=2)
+# one extra query to convert index to native pandas when creating the snowpark pandas dataframe
+@sql_count_checker(query_count=2, join_count=2)
 def test_dataframe_where_with_dataframe_cond_single_index_different_names_2():
     data = [1, 2, 3]
     cond = [False, True, False]
@@ -701,7 +703,8 @@ def test_dataframe_where_with_duplicated_index_aligned(cond_frame, other):
         )
 
 
-@sql_count_checker(query_count=1, join_count=2)
+# 3 extra queries to convert index to native pandas when creating the 3 snowpark pandas dataframe
+@sql_count_checker(query_count=4, join_count=2)
 def test_dataframe_where_with_duplicated_index_unaligned():
     data = [3, 4, 5, 2]
     df_index = pd.Index([2, 1, 2, 3], name="index")